+++ visionatrix/install_update/update.py
@@ -6,7 +6,7 @@ from subprocess import CalledProcessError, check_call
 
 from packaging.version import Version
 
-from .. import _version, comfyui, options
+from .. import _version, comfyui_wrapper, options
 from ..flows import get_available_flows, get_installed_flows, install_custom_flow
 from .custom_nodes import update_base_custom_nodes
 from .install import create_nodes_stuff
@@ -40,7 +40,7 @@ def update() -> None:
         create_nodes_stuff()
         logging.info("Updating custom nodes..")
         update_base_custom_nodes()
-    comfyui.load(None)
+    comfyui_wrapper.load(None)
     logging.info("Updating flows..")
     avail_flows_comfy = {}
     avail_flows = get_available_flows(avail_flows_comfy)
